requirements.txt
matplotlib
pandas
Pillow
streamlit
mysql.connector
openpyxl
streamlit-authenticator